# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Supermarket Store Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: warehouseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List inventory movements
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Movements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      summary: Create inventory movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"

  # SUPPLIERS
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leaves]
      summary: Create leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/status:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      summary: Approve/reject leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # PAYROLL
  /payroll:
    get:
      operationId: listPayroll
      tags: [payroll]
      summary: List payroll records
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payroll records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRecord"
    post:
      operationId: createPayrollRecord
      tags: [payroll]
      summary: Create payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRecord"
  /payroll/{id}:
    get:
      operationId: getPayrollRecord
      tags: [payroll]
      summary: Get payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRecord"
  /payroll/process:
    post:
      operationId: processPayroll
      tags: [payroll]
      summary: Process monthly payroll for all employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollProcessInput"
      responses:
        "200":
          description: Payroll processed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRecord"

  # EXPENSES
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [id, productId, productName, quantity, unit]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: number
        reservedQty:
          type: number
        damagedQty:
          type: number
        unit:
          type: string
        warehouseId:
          type: ["integer", "null"]
        warehouseName:
          type: ["string", "null"]
        updatedAt:
          type: string
    InventoryMovement:
      type: object
      required: [id, productId, type, quantity, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [stock_in, stock_out, transfer, adjustment, damaged, returned]
        quantity:
          type: number
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    InventoryMovementInput:
      type: object
      required: [productId, type, quantity]
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [stock_in, stock_out, transfer, adjustment, damaged, returned]
        quantity:
          type: number
        reference:
          type: string
        notes:
          type: string
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
...
    LeaveRequest:
      type: object
      required: [id, employeeId, leaveType, startDate, endDate, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveInput:
      type: object
      required: [employeeId, leaveType, startDate, endDate]
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    PayrollRecord:
      type: object
      required: [id, employeeId, month, year, basicSalary, netPay, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        bonus:
          type: number
        allowances:
          type: number
        deductions:
          type: number
        tax:
          type: number
        netPay:
          type: number
        status:
          type: string
          enum: [pending, processed, paid]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PayrollInput:
      type: object
      required: [employeeId, month, year, basicSalary]
      properties:
        employeeId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        bonus:
          type: number
        allowances:
          type: number
        deductions:
          type: number
        tax:
          type: number
    PayrollProcessInput:
      type: object
      required: [month, year]
      properties:
        month:
          type: integer
        year:
          type: integer
    Expense:
      type: object
      required: [id, title, category, amount, date]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [rent, electricity, water, maintenance, internet, transportation, salaries, supplies, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: ["string", "null"]
        paymentMethod:
          type: string
        status:
          type: string
          enum: [pending, approved, paid]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [title, category, amount, date]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [rent, electricity, water, maintenance, internet, transportation, salaries, supplies, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [rent, electricity, water, maintenance, internet, transportation, salaries, supplies, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
          enum: [pending, approved, paid]
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [low_stock, expiring_product, payroll_due, purchase_approval, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
...