# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Supermarket Store Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseUpdateCategory } from './expenseUpdateCategory';
import type { ExpenseUpdateStatus } from './expenseUpdateStatus';

export interface ExpenseUpdate {
  title?: string;
  category?: ExpenseUpdateCategory;
  amount?: number;
  date?: string;
  description?: string;
  paymentMethod?: string;
  status?: ExpenseUpdateStatus;
}